Fix: Pass page linear
This bug was due to dirty bit.
In previous implementation, I just use upage dirty bit.
I tried to fix this bug for 4 hours, and I finally found dirty wrong.
At that time, I cannot understand why it is not working.
I did almost all things, and i found that i should use kpage dirty.
In previous commit, it already fixed.
It is because i have to commit this to test on server.
So this commit is.. kind of useless commit haha.
Anyway, I FIXED IT!!!!!!!!!!!!!!!!!!.
JONNA PPAK CHUE.